refactor: move use-run-search to common
Refs: CO-1741
