+++ src/views/app/detail-panel/preview/parts/preview-header.tsx
@@ -52,11 +52,11 @@ import MessageContactsList from './message-contact-list';
 import OnBehalfOfDisplayer from './on-behalf-of-displayer';
 import { ParticipantRole } from '../../../../../carbonio-ui-commons/constants/participants';
 import { ZIMBRA_STANDARD_COLORS } from '../../../../../carbonio-ui-commons/constants/utils';
+import { useRunSearchIntegration } from '../../../../../carbonio-ui-commons/integrations/search/use-run-search';
 import { useTags } from '../../../../../carbonio-ui-commons/store/zustand/tags';
 import { Tag } from '../../../../../carbonio-ui-commons/types/tags';
 import { getTimeLabel, participantToString } from '../../../../../commons/utils';
 import { getNoIdentityPlaceholder } from '../../../../../helpers/identities';
-import { useRunSearchIntegration } from '../../../../../integrations/search/use-run-search';
 import { retrieveAttachmentsType } from '../../../../../store/editor-slice-utils';
 import type { MailMessage } from '../../../../../types';
 import { useTagExist } from '../../../../../ui-actions/tag-actions';
